+++ src/components/CartProductsList/index.tsx
@@ -13,11 +13,11 @@ import RemoveCircleIcon from "@material-ui/icons/RemoveCircle";
 import { graphql } from "babel-plugin-relay/macro";
 import { createFragmentContainer } from "react-relay";
 import { CartProductsList_shoppingCart } from "./__generated__/CartProductsList_shoppingCart.graphql";
-import { formatCurrency } from "../../../utils/formaters";
-import { uuidVersion4Generator } from "../../../utils/idGenerators";
-import { updateProduct } from "../../../relay/mutations/UpdateProduct";
-import { updateCartProduct } from "../../../relay/mutations/UpdateCartProduct";
-import { MySnackbarContentWrapper } from "../../SnackBar";
+import { formatCurrency } from "../../utils/formaters";
+import { uuidVersion4Generator } from "../../utils/idGenerators";
+import { updateProduct } from "../../relay/mutations/UpdateProduct";
+import { updateCartProduct } from "../../relay/mutations/UpdateCartProduct";
+import { MySnackbarContentWrapper } from "../SnackBar";
 import { Link } from "react-router-dom";
 
 const useStyles = makeStyles((theme: Theme) =>
